Renamed RecipeDisplay folder to recipeBrief
This is aligned with our naming convention (lowercase folders) and also convention of co-locating a single component and component style file in one folder
